requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
django-environ==0.8.1
djangorestframework==3.14.0
Markdown==3.4.1
psycopg2==2.9.3
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
drf-yasg==1.21.4 # Libreary for swagger and redoc
django-cors-headers==3.11.0
django-simple-history==3.0.0
pillow==9.0.1
# Heroku
whitenoise==6.2.0
django-heroku==0.3.1